Fixed some bugs.
