+++ .idea/.gitignore
@@ -1,0 +1,2 @@
+# Default ignored files
+/workspace.xml
