+++ requirements.txt
@@ -1,3 +1,4 @@
 PyMySQL==0.9.3
 PyQt5==5.12
 SQLAlchemy==1.2.17
+pandas==0.23.4
