+++ resources/sql_comparator_backlog.txt
@@ -10,6 +10,7 @@ than some hardcoded value, I should decrease comparing step.
 6. Fix sendmail.
 7. Add special search lineedits to modal windows like "skip_tables".
 8. Only first error flag not works.
+9. On progress window fix "Processing"-string. It should be made with different layout to not cut tablenames.
 
 Epics:
 0. Support dataframe comparing.
